+++ BAVCL/Core/Enums/AsType.cs
@@ -1,0 +1,8 @@
+namespace BAVCL.Core;
+
+public enum AsType
+{
+    Vector = 0,
+    Vector3 = 1,
+    Array = 2,
+}
