+++ BAVCL/Core/Enums/Axis.cs
@@ -1,0 +1,7 @@
+namespace BAVCL.Core;
+
+public enum Axis
+{
+    Row = 0,
+    Column = 1,
+}
