--- BAVCL/Core/Enums/Enums.cs
@@ -1,24 +1,0 @@
-namespace BAVCL.Core;
-
-public enum FileTypes
-{
-    CSV = 0,
-    TXT = 1,
-    // FITS
-    // JSON
-    // XML
-
-}
-
-public enum Axis
-{
-    Row = 0,
-    Column = 1,
-}
-
-public enum AsType
-{
-    Vector = 0,
-    Vector3 = 1,
-    Array = 2,
-}
