--- BAVCL/Core/Table/Table.cs
@@ -1,5 +1,0 @@
-namespace BAVCL.Core;
-
-class Table
-{
-}
