+++ BAVCL/Exceptions/KernelNotCompiledException.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace BAVCL.Core.Exceptions;
+namespace BAVCL.Exceptions;
 
 public class KernelNotCompiledException(string kernelName) : Exception(
 $"Kernel: '{kernelName}' not compiled. Ensure that it has been compiled before use.")
