--- BAVCL/Geometric/Vector3/Copy.cs
@@ -1,13 +1,0 @@
-namespace BAVCL.Geometric;
-
-public sealed partial class Vector3 : VectorBase<float>
-{
-    public Vector3 Copy()
-    {
-        if (_id != 0)
-        {
-            return new Vector3(Gpu, Pull());
-        }
-        return new Vector3(Gpu, ToArray());
-    }
-}
