--- BAVCL/Geometric/Vector3/SetValue.cs
@@ -1,15 +1,0 @@
-using BAVCL.Geometric.Enums;
-using System;
-
-namespace BAVCL.Geometric;
-
-public partial class Vector3
-{
-    public void SetAt(int row, Coord coord, float value)
-    {
-        if (row < 0 || row > RowCount())
-            throw new IndexOutOfRangeException();
-
-        SetAt(row, (int)coord, value);
-    }
-}
