+++ BAVCL/Modules/Geometric/Types/Coord.cs
