+++ BAVCL/Modules/Geometric/Types/Vertex.cs
