+++ BAVCL/Modules/IO/Enums/FileType.cs
@@ -1,0 +1,11 @@
+namespace BAVCL.Modules.IO.Enums;
+
+public enum FileTypes
+{
+    CSV = 0,
+    TXT = 1,
+    // FITS
+    // JSON
+    // XML
+
+}
