+++ BAVCL/Types/Vector.cs
